New/desafios

* Adicionado novas missões.

* Banner de promoções.

* Implementação de alguns desafios para validar o fluxo geral.

* Desafios implementados.

* Layout definitivo do ranking de guerra.

* CHG: Altarado estrutura de migrations para a aplicação ficar automática e sob demanda.

* NEW: Criação da estrutura para a meta de doações.

* Portas do websocker e do servidor passadas por argumento no backend.

* Baixando imagem das premiações e gerando desafios apenas caso não tenha sido gerado ainda.

* Alterado versão dos assets para 2.2.

+++ implementacao/server/migrations/database_update1.py
@@ -1,10 +1,21 @@
 import sqlite3
+import sys
 
-con = sqlite3.connect('war.db')
+wardb = 'war.db'
+if len(sys.argv) > 1:
+    wardb = sys.argv[1]
 
+con = sqlite3.connect(wardb)
 with con:
     cur = con.cursor()
-    cur.execute("DROP TABLE IF EXISTS Usuarios")
-    cur.execute("CREATE TABLE Usuarios(id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL, nome TEXT, senha TEXT, email TEXT)")
-    cur.execute("INSERT INTO Versao(sistema, versao) VALUES ('db', '1')")
-    con.commit()
+    row = cur.execute("SELECT versao FROM Versao WHERE sistema = 'db' ORDER BY versao DESC LIMIT 1").fetchone()
+    if row and row[0] == '0':
+        print('Aplicando Patch database_update1...')
+        cur.execute("DROP TABLE IF EXISTS Usuarios")
+        cur.execute(
+            "CREATE TABLE Usuarios(id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL, nome TEXT, senha TEXT, email TEXT)")
+        cur.execute("INSERT INTO Versao(sistema, versao) VALUES ('db', '1')")
+        con.commit()
+        print('Patch database_update1 aplicado com sucesso.')
+    else:
+        print('Patch database_update1 já foi aplicado anteriormente.')
