+++ implementacao/server/migrations/database_update5.py
@@ -1,10 +1,21 @@
 import sqlite3
+import sys
 
-con = sqlite3.connect('war.db')
+wardb = 'war.db'
+if len(sys.argv) > 1:
+    wardb = sys.argv[1]
 
+con = sqlite3.connect(wardb)
 with con:
     cur = con.cursor()
-    cur.execute("DROP TABLE IF EXISTS RecuperacaoSenha")
-    cur.execute("CREATE TABLE IF NOT EXISTS RecuperacaoSenha(email TEXT PRIMARY KEY, codigo TEXT, data DATETIME DEFAULT current_timestamp)")
-    cur.execute("INSERT INTO Versao(sistema, versao) VALUES ('db', '5')")
-    con.commit()
+    row = cur.execute("SELECT versao FROM Versao WHERE sistema = 'db' ORDER BY versao DESC LIMIT 1").fetchone()
+    if row and row[0] == '4':
+        print('Aplicando Patch database_update5...')
+        cur.execute("DROP TABLE IF EXISTS RecuperacaoSenha")
+        cur.execute(
+            "CREATE TABLE IF NOT EXISTS RecuperacaoSenha(email TEXT PRIMARY KEY, codigo TEXT, data DATETIME DEFAULT current_timestamp)")
+        cur.execute("INSERT INTO Versao(sistema, versao) VALUES ('db', '5')")
+        con.commit()
+        print('Patch database_update5 aplicado com sucesso.')
+    else:
+        print('Patch database_update5 já foi aplicado anteriormente.')
